safari-600.7-branch/LayoutTests/inspector-obsolete/http-tests/console-xhr-logging-async-expected.txt

CONSOLE MESSAGE: line 6: XHR loaded: 1
CONSOLE MESSAGE: line 6: XHR loaded: 2
Tests that XMLHttpRequest Logging works when Enabled and doesn't show logs when Disabled for asynchronous XHRs.

Bug 79229
XHR finished loading: "http://127.0.0.1:8000/inspector/resources/xhr-exists.html". [native code]:1
XHR loaded: 1 network-test.js:6
XHR loaded: 2 network-test.js:6
